Update kube-scheduler configuration
Update kube-scheduler configuration to be compatible with k8s v1.22. The
scheduler configuration set via commandline flags will be removed. Most
of them were deprecated in k8s v1.18.

The configuration is set via KubeSchedulerConfiguration v1beta1 API in
the form of a config file.
The scheduler name is set by creating scheduler profiles. A scheduler
can have multiple scheduler profiles. A scheduler profile can have a
set of plugins to perform scheduling. These plugins are based on the new
scheduling framework. When no plugin is set, the default scheduler
filtering and scoring is used.
The scheduler extender policy is set at the root of the
KubeSchedulerConfiguration. These scheduler extender policies are
applied to all the scheduler profiles in the scheduler.
The leader election configuration is also set at the root of the
KubeSchedulerConfiguration.